server/requirements.txt
Django==3.0.8
psycopg2-binary==2.8.5 # работа с базой данных postgres

dj-database-url==0.5.0 # передача настроек БД из env окружения
gunicorn==20.0.4 # http сервер, который свзявает nginx с django приложением

ipdb==0.13.3 # отладка программ
ipython==7.12.0 # приятная работа в shell
django-debug-toolbar==2.2 # панель для отладки запросов в бд

Pillow==7.2.0 # для работы с изображениями
djangorestframework==3.11.0 # API
django-url-filter==0.3.15 # улучшенные фильтры для DRF

yandex-geocoder==2.0.0 # Get address/coordinates via Yandex geocoder
beautifulsoup4==4.9.1 # bs4 для парсинга ингредиентов
lxml==4.5.2 # для работы bs4